Remove _posix file (not necessary)
